refactor(Client): :recycle: rename api.ts to webAppApi.ts
